pyproject.toml
[project]
name = "fastapi-app"
version = "0.1.0"
description = "Base template FastAPI + Tortoise + Decouple"
authors = [
    {name = "jonathan001",email = "jonsecu001@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[all] (>=0.115.14,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "tortoise-orm (>=0.25.0,<0.26.0)",
    "python-decouple (>=3.8,<4.0)",
    "pydantic[email] (>=2.11.5,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^1.0.0"
httpx = "^0.24.0"

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"

